+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - gen_class.Yum: Added python version check so class will only work in Python 2.7, yum==3.4.3 does not work in Python 3.
 
 ### Changed
+- gen_libs.file_search_cnt: Refactored function to work in Python 3.
 - gen_class.ArgParser: get_args_keys, arg_wildcard: Converted a dict.keys() to a list.
 - gen_class.Daemon: Set class to use the superclass object.
 - gen_class: TimeFormat.\_\_init\_\_, ProgressBar.update: Changed division to floor division.